lambda_functions/telegram_bot/requirements.txt
langchain==0.2.10
langchain-aws==0.1.12
langchain-community==0.2.9
python-telegram-bot==21.4
